requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
plotly==5.15.0
tensorflow==2.13.0
python-dotenv==1.0.0
scipy==1.11.1
